backend/requirements.txt
# FastAPI & ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Gemini API
google-generativeai==0.3.2

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTML/XML parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

